Satisfy CodeFactor
